backend/workers/summarizer_worker/requirements.txt
# Inherit from parent
-r ../../requirements.txt

# Additional dependencies
PyPDF2==3.0.1
langchain==0.0.340
langchain-google-vertexai==0.0.6
